+++ src/Region/RegionHandler.h
@@ -9,6 +9,7 @@
 #ifndef CARTA_SRC_REGION_REGIONHANDLER_H_
 #define CARTA_SRC_REGION_REGIONHANDLER_H_
 
+#include <mutex>
 #include <vector>
 
 #include "Cache/RequirementsCache.h"
@@ -18,6 +19,7 @@
 #include "ImageGenerators/PvPreviewCut.h"
 #include "Region.h"
 #include "RegionAnalysis/RegionHistogram.h"
+#include "RegionAnalysis/RegionSpatialProfile.h"
 #include "RegionAnalysis/RegionStatistics.h"
 
 namespace carta {
@@ -44,8 +46,8 @@ public:
     // Requirements
     bool SetHistogramRequirements(
         int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::HistogramConfig>& configs);
-    bool SetSpatialRequirements(int region_id, int file_id, std::shared_ptr<Frame> frame,
-        const std::vector<CARTA::SetSpatialRequirements_SpatialConfig>& spatial_profiles);
+    bool SetSpatialRequirements(
+        int region_id, int file_id, std::shared_ptr<Frame> frame, const std::vector<CARTA::SetSpatialRequirements_SpatialConfig>& configs);
     bool SetSpectralRequirements(int region_id, int file_id, std::shared_ptr<Frame> frame,
         const std::vector<CARTA::SetSpectralRequirements_SpectralConfig>& configs);
     bool SetStatsRequirements(
@@ -56,8 +58,7 @@ public:
     bool FillSpectralProfileData(
         std::function<void(CARTA::SpectralProfileData profile_data)> cb, int region_id, int file_id, bool stokes_changed);
     bool FillRegionStatsData(std::function<void(CARTA::RegionStatsData stats_data)> cb, int region_id, int file_id);
-    bool FillPointSpatialProfileData(int file_id, int region_id, std::vector<CARTA::SpatialProfileData>& spatial_data_vec);
-    bool FillLineSpatialProfileData(int file_id, int region_id, std::function<void(CARTA::SpatialProfileData profile_data)> cb);
+    bool FillSpatialProfileData(std::function<void(CARTA::SpatialProfileData profile_data)> cb, int file_id, int region_id);
 
     // Calculate moments
     bool CalculateMoments(int file_id, int region_id, const std::shared_ptr<Frame>& frame, GeneratorProgressCallback progress_callback,
@@ -68,12 +69,8 @@ public:
     bool IsLineRegion(int region_id);
     bool IsClosedRegion(int region_id);
 
-    // Spatial Requirements
-    std::vector<int> GetSpatialReqRegionsForFile(int file_id);
-    std::vector<int> GetSpatialReqFilesForRegion(int region_id);
-
     // Generate PV image or preview image
-    bool CalculatePvImage(const CARTA::PvRequest& pv_request, std::shared_ptr<Frame>& frame, GeneratorProgressCallback progress_callback,
+    bool CalculatePvImage(const CARTA::PvRequest& pv_request, std::shared_ptr<Frame> frame, GeneratorProgressCallback progress_callback,
         CARTA::PvResponse& pv_response, GeneratedImage& pv_image);
     // Update PV preview image when PV cut region (region_id) changes
     bool UpdatePvPreviewRegion(int region_id, RegionState& region_state);
@@ -101,17 +98,10 @@ private:
     // Requirements helpers
     bool HasSpectralRequirements(
         int region_id, int file_id, const std::string& coordinate, const std::vector<CARTA::StatsType>& required_stats);
-    bool HasSpatialRequirements(int region_id, int file_id, const std::string& coordinate, int width);
     void UpdateNewSpectralRequirements(int region_id);
     void RemoveRegionRequirementsCache(int region_id);
     void RemoveFileRequirementsCache(int file_id);
     void ClearRegionCache(int region_id);
-
-    // Apply region to image
-    std::shared_ptr<casacore::LCRegion> ApplyRegionToFile(
-        int region_id, int file_id, const StokesSource& stokes_source = StokesSource(), bool report_error = true);
-    bool ApplyRegionToFile(int region_id, int file_id, const AxisRange& z_range, int stokes, std::shared_ptr<casacore::LCRegion> lc_region,
-        StokesRegion& stokes_region);
 
     // Data stream helpers
     bool GetRegionSpectralData(int region_id, int file_id, const AxisRange& z_range, std::string& coordinate, int stokes_index,
@@ -120,14 +110,13 @@ private:
     bool GetLineSpatialData(int file_id, int region_id, const std::string& coordinate, int stokes_index, int width,
         const std::function<void(std::vector<float>&, casacore::Quantity&)>& spatial_profile_callback);
 
-    // Generate box regions to approximate a line with a width, and get mean of each box for z-range.
-    // Used for pv generator and spatial profiles (reverse only applies to PV).
-    bool GetLineProfiles(int file_id, int region_id, int width, const AxisRange& z_range, bool per_z, int stokes_index,
-        const std::string& coordinate, std::function<void(float)>& progress_callback, casacore::Matrix<float>& profiles,
-        casacore::Quantity& increment, bool& cancelled, std::string& message, bool reverse = false);
+    // Generate box regions to approximate a line with a width, and get mean of each box for z-range for PV image.
+    bool GetLineProfiles(int file_id, int region_id, int width, const AxisRange& z_range, int stokes_index, const std::string& coordinate,
+        std::function<void(float)>& progress_callback, casacore::Matrix<float>& profiles, casacore::Quantity& increment, bool& cancelled,
+        std::string& message, bool reverse = false);
     bool CancelLineProfiles(int region_id, int file_id, RegionState& region_state);
-    casacore::Vector<float> GetTemporaryRegionProfile(int region_idx, int file_id, RegionState& region_state,
-        std::shared_ptr<casacore::CoordinateSystem> csys, bool per_z, const AxisRange& z_range, int stokes_index, double& num_pixels);
+    casacore::Vector<float> GetTemporaryRegionProfile(int file_id, RegionState& region_state,
+        std::shared_ptr<casacore::CoordinateSystem> csys, const AxisRange& z_range, int stokes_index, double& num_pixels);
 
     // Get computed stokes profiles for a region
     using ProfilesMap = std::map<CARTA::StatsType, std::vector<double>>;
@@ -144,7 +133,7 @@ private:
     void CombineStokes(ProfilesMap& profiles_out, const ProfilesMap& profiles_other, const std::function<double(double, double)>& func);
     bool IsValid(double a, double b);
     bool GetComputedStokesProfiles(
-        ProfilesMap& profiles, int stokes, const std::function<bool(ProfilesMap&, std::string)>& get_profiles_data);
+        ProfilesMap& profiles, int stokes_index, const std::function<bool(ProfilesMap&, std::string)>& get_profiles_data);
 
     // PV generator
     bool CalculatePvImage(int file_id, int region_id, int width, AxisRange& spectral_range, bool reverse, bool keep,
@@ -168,19 +157,14 @@ private:
 
     // Region analysis
     std::unordered_map<int, std::unique_ptr<RegionHistogram>> _region_histograms;
+    std::unordered_map<int, std::shared_ptr<RegionSpatialProfile>> _region_spatial_profiles;
     std::unordered_map<int, std::unique_ptr<RegionStatistics>> _region_statistics;
-
-    // Requirements; ConfigId key contains file, region
-    std::unordered_map<ConfigId, RegionSpectralConfig, ConfigIdHash> _spectral_req;
-    std::unordered_map<ConfigId, std::vector<CARTA::SetSpatialRequirements_SpatialConfig>, ConfigIdHash> _spatial_req;
-    // Lock to add/remove requirements
     std::mutex _spatial_mutex;
-    std::mutex _spectral_mutex;
 
-    // Cache; CacheId key contains file, region, stokes, (optional) z index
+    // Spectral profile requirements, cache, and list of supported stats
+    std::unordered_map<ConfigId, RegionSpectralConfig, ConfigIdHash> _spectral_req;
     std::unordered_map<CacheId, SpectralCache, CacheIdHash> _spectral_cache;
-
-    // Spectral profiles to calculate with ImageStatistics.
+    std::mutex _spectral_mutex;
     std::vector<CARTA::StatsType> _spectral_stats = {CARTA::StatsType::Sum, CARTA::StatsType::FluxDensity, CARTA::StatsType::Mean,
         CARTA::StatsType::RMS, CARTA::StatsType::Sigma, CARTA::StatsType::SumSq, CARTA::StatsType::Min, CARTA::StatsType::Max,
         CARTA::StatsType::Extrema, CARTA::StatsType::NumPixels};