+++ test/TestRegionSpatialProfiles.cc
@@ -64,8 +64,19 @@ public:
         return region_handler.SetRegion(region_id, region_state, csys);
     }
 
+    static bool SetRectangleRegion(carta::RegionHandler& region_handler, int file_id, int& region_id, const std::vector<float>& points,
+        std::shared_ptr<casacore::CoordinateSystem> csys) {
+        std::vector<CARTA::Point> control_points;
+        for (auto i = 0; i < points.size(); i += 2) {
+            control_points.push_back(Message::Point(points[i], points[i + 1]));
+        }
+        CARTA::RegionType region_type(CARTA::RECTANGLE);
+        RegionState region_state(file_id, region_type, control_points, 0.0);
+        return region_handler.SetRegion(region_id, region_state, csys);
+    }
+
     static bool RegionSpatialProfile(const std::string& image_path, const std::vector<float>& endpoints,
-        const std::vector<CARTA::SetSpatialRequirements_SpatialConfig>& spatial_reqs, CARTA::SpatialProfileData& spatial_profile,
+        const std::vector<CARTA::SetSpatialRequirements_SpatialConfig>& spatial_reqs, CARTA::SpatialProfileData& spatial_profile_message,
         bool is_annotation = false) {
         std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(image_path));
         std::shared_ptr<Frame> frame(new Frame(0, loader, "0"));
@@ -74,6 +85,7 @@ public:
         // Set line region
         int file_id(0), region_id(-1);
         auto csys = frame->CoordinateSystem();
+
         if (!SetRegion(region_handler, file_id, region_id, endpoints, csys, is_annotation)) {
             return false;
         }
@@ -84,17 +96,8 @@ public:
         }
 
         // Get spatial profiles
-        if (endpoints.size() == 2) {
-            std::vector<CARTA::SpatialProfileData> profiles;
-            bool ok = region_handler.FillPointSpatialProfileData(file_id, region_id, profiles);
-            if (ok) {
-                spatial_profile = profiles[0];
-            }
-            return ok;
-        } else {
-            return region_handler.FillLineSpatialProfileData(
-                file_id, region_id, [&](CARTA::SpatialProfileData profile_data) { spatial_profile = profile_data; });
-        }
+        return region_handler.FillSpatialProfileData(
+            [&](CARTA::SpatialProfileData profile_message) { spatial_profile_message = profile_message; }, file_id, region_id);
     }
 
     static std::vector<float> ProfileValues(CARTA::SpatialProfile& profile) {
@@ -110,18 +113,18 @@ public:
 
     static void TestAveragingWidthRange(int width, bool expected_width_range) {
         std::string image_path = FileFinder::FitsImagePath("noise_3d.fits");
+        int file_id(0), region_id(-1);
         std::vector<float> endpoints = {0.0, 0.0, 9.0, 9.0};
         int start(0), end(0), mip(0);
-        std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
-            RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
-        CARTA::SpatialProfileData spatial_profile;
+        std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
+        CARTA::SpatialProfileData spatial_profile_message;
 
         if (expected_width_range) {
-            ASSERT_TRUE(RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile));
-            ASSERT_EQ(spatial_profile.profiles_size(), 1);
+            ASSERT_TRUE(RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message));
+            ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
         } else {
-            ASSERT_FALSE(RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile));
-            ASSERT_EQ(spatial_profile.profiles_size(), 0);
+            ASSERT_FALSE(RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message));
+            ASSERT_EQ(spatial_profile_message.profiles_size(), 0);
         }
     }
 };
@@ -147,15 +150,15 @@ TEST_F(RegionSpatialProfileTest, TestSpatialRequirements) {
     ok = region_handler.SetSpatialRequirements(region_id, file_id, frame, spatial_reqs);
     ASSERT_TRUE(ok);
 
-    // Get regions with spatial requirements for file id
-    auto region_ids = region_handler.GetSpatialReqRegionsForFile(file_id);
-    ASSERT_EQ(region_ids.size(), 1);
-    ASSERT_EQ(region_ids[0], region_id);
+    // Set rectangle region
+    region_id = -1;
+    endpoints = {5.0, 5.0, 3.0, 2.0};
+    ok = SetRectangleRegion(region_handler, file_id, region_id, endpoints, csys);
+    ASSERT_TRUE(ok);
 
-    // Get files with spatial requirements for region id
-    auto file_ids = region_handler.GetSpatialReqFilesForRegion(region_id);
-    ASSERT_EQ(file_ids.size(), 1);
-    ASSERT_EQ(file_ids[0], file_id);
+    // Set spatial requirements for invalid region type
+    ok = region_handler.SetSpatialRequirements(region_id, file_id, frame, spatial_reqs);
+    ASSERT_FALSE(ok);
 }
 
 TEST_F(RegionSpatialProfileTest, FitsLineProfile) {
@@ -165,13 +168,11 @@ TEST_F(RegionSpatialProfileTest, FitsLineProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
-
-    // Check file/region ids for requirements
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 }
 
 TEST_F(RegionSpatialProfileTest, Hdf5LineProfile) {
@@ -181,11 +182,11 @@ TEST_F(RegionSpatialProfileTest, Hdf5LineProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 }
 
 TEST_F(RegionSpatialProfileTest, FitsHorizontalCutProfile) {
@@ -195,14 +196,14 @@ TEST_F(RegionSpatialProfileTest, FitsHorizontalCutProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 
     // Profile data
-    auto profile = spatial_profile.profiles(0);
+    auto profile = spatial_profile_message.profiles(0);
     std::vector<float> profile_data = ProfileValues(profile);
     EXPECT_EQ(profile_data.size(), 9);
 
@@ -220,14 +221,14 @@ TEST_F(RegionSpatialProfileTest, FitsVerticalCutProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("y", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 
     // Profile data
-    auto profile = spatial_profile.profiles(0);
+    auto profile = spatial_profile_message.profiles(0);
     std::vector<float> profile_data = ProfileValues(profile);
     EXPECT_EQ(profile_data.size(), 9);
 
@@ -242,17 +243,15 @@ TEST_F(RegionSpatialProfileTest, FitsPolylineProfile) {
     std::string image_path = FileFinder::FitsImagePath("noise_3d.fits");
     std::vector<float> endpoints = {1.0, 1.0, 9.0, 1.0, 9.0, 5.0};
     int start(0), end(0), mip(0), width(1);
-    std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
-        RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
-    std::vector<CARTA::SpatialProfileData> spatial_profiles;
+    std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 
     // Profile data
-    auto profile = spatial_profile.profiles(0);
+    auto profile = spatial_profile_message.profiles(0);
     std::vector<float> profile_data = ProfileValues(profile);
     EXPECT_EQ(profile_data.size(), 13);
 
@@ -284,10 +283,9 @@ TEST_F(RegionSpatialProfileTest, FitsAnnotationLineProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
+    CARTA::SpatialProfileData spatial_profile_message;
     bool is_annotation(true);
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile, is_annotation);
-
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message, is_annotation);
     ASSERT_FALSE(ok);
 }
 
@@ -296,16 +294,16 @@ TEST_F(RegionSpatialProfileTest, FitsPointProfile) {
     std::vector<float> endpoints = {0.0, 0.0};
     int start(0), end(0), mip(0), width(1);
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
-        RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
+        RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width), RegionSpatialProfileTest::SpatialConfig("y", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 2);
 
     // Profile data for 10 x 10 image frame
-    auto profile = spatial_profile.profiles(0);
+    auto profile = spatial_profile_message.profiles(0);
     std::vector<float> profile_data = ProfileValues(profile);
     EXPECT_EQ(profile_data.size(), 10);
 
@@ -323,14 +321,14 @@ TEST_F(RegionSpatialProfileTest, Hdf5PointProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, points, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, points, spatial_reqs, spatial_profile_message);
 
     ASSERT_TRUE(ok);
-    ASSERT_EQ(spatial_profile.profiles_size(), 1);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 
     // Profile data for 10 x 10 image frame
-    auto profile = spatial_profile.profiles(0);
+    auto profile = spatial_profile_message.profiles(0);
     std::vector<float> profile_data = ProfileValues(profile);
     EXPECT_EQ(profile_data.size(), 10);
 
@@ -348,8 +346,8 @@ TEST_F(RegionSpatialProfileTest, FitsPointProfileOutsideImage) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
-    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile);
+    CARTA::SpatialProfileData spatial_profile_message;
+    bool ok = RegionSpatialProfile(image_path, endpoints, spatial_reqs, spatial_profile_message);
 
     ASSERT_FALSE(ok);
 }
@@ -361,9 +359,48 @@ TEST_F(RegionSpatialProfileTest, FitsAnnotationPointProfile) {
     std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {
         RegionSpatialProfileTest::SpatialConfig("x", start, end, mip, width)};
 
-    CARTA::SpatialProfileData spatial_profile;
+    CARTA::SpatialProfileData spatial_profile_message;
     bool is_annotation(true);
-    bool ok = RegionSpatialProfile(image_path, point, spatial_reqs, spatial_profile, is_annotation);
+    bool ok = RegionSpatialProfile(image_path, point, spatial_reqs, spatial_profile_message, is_annotation);
 
     ASSERT_FALSE(ok);
+}
+
+TEST_F(RegionSpatialProfileTest, FitsMovePointProfile) {
+    std::string image_path = FileFinder::FitsImagePath("noise_3d.fits");
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(image_path));
+    std::shared_ptr<Frame> frame(new Frame(0, loader, "0"));
+
+    // Set point region
+    carta::RegionHandler region_handler;
+    int file_id(0), region_id(-1);
+    std::vector<float> endpoints = {0.0, 0.0};
+    auto csys = frame->CoordinateSystem();
+    bool is_annotation(false);
+    bool ok = SetRegion(region_handler, file_id, region_id, endpoints, csys, is_annotation);
+    ASSERT_TRUE(ok);
+
+    // Set spatial requirements
+    int start(0), end(0), mip(0), width(3);
+    std::vector<CARTA::SetSpatialRequirements_SpatialConfig> spatial_reqs = {Message::SpatialConfig("x", start, end, mip, width)};
+    ok = region_handler.SetSpatialRequirements(region_id, file_id, frame, spatial_reqs);
+    ASSERT_TRUE(ok);
+
+    // Get spatial profile
+    CARTA::SpatialProfileData spatial_profile_message;
+    ok = region_handler.FillSpatialProfileData(
+        [&](CARTA::SpatialProfileData profile_message) { spatial_profile_message = profile_message; }, file_id, region_id);
+    ASSERT_TRUE(ok);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
+
+    // Move point
+    endpoints = {9.0, 9.0};
+    ok = SetRegion(region_handler, file_id, region_id, endpoints, csys, is_annotation);
+    ASSERT_TRUE(ok);
+
+    // Get spatial profile
+    ok = region_handler.FillSpatialProfileData(
+        [&](CARTA::SpatialProfileData profile_message) { spatial_profile_message = profile_message; }, file_id, region_id);
+    ASSERT_TRUE(ok);
+    ASSERT_EQ(spatial_profile_message.profiles_size(), 1);
 }